requirements.txt
Flask==3.1.1
python-dotenv==1.0.1
google-genai==1.21.0
google-generativeai==0.8.5
google-cloud-storage==3.1.1
moviepy==1.0.3
requests==2.31.0
numpy==1.26.4
Jinja2==3.1.4
gunicorn==22.0.0
openai==1.30.1
imageio==2.34.0
imageio-ffmpeg==0.4.9
decorator==4.4.2
tqdm==4.66.4
# transformers==4.41.2
# torch==2.7.1
gradio_client>=0.6.1
pychorus==0.1.0
pydub==0.25.1
